Added separate line for debug flag in Makefile. Reduced the number of casts in the ADC_ISR, which should decrease the time spent in the ISR. Now casts to uint32_t when we decide to print.

+++ report.c
@@ -384,7 +384,7 @@ void report_voltage()
   uint8_t i;
   printPgmString(PSTR("|"));
   for (i = 0; i<VOLT_NUM; i++){
-    printInteger(voltage_result[i]);
+    printInteger((uint32_t)voltage_result[i]);
     if (i<VOLT_NUM-1)
       printPgmString(PSTR(","));
   }
